move public utils into vector db base class
