+++ ragna/__init__.py
@@ -8,5 +8,5 @@ except ModuleNotFoundError:
 
     __version__ = "UNKNOWN"
 
-from . import assistants, core, source_storages, utils
+from . import assistants, core, source_storages
 from .core import Config, Rag
