--- ragna/utils/__init__.py
@@ -1,7 +1,0 @@
-from ragna._utils import fix_module  # usort: skip
-
-from ._chunk import Chunk, Tokenizer, chunk_pages
-from ._misc import page_numbers_to_str, take_sources_up_to_max_tokens
-
-fix_module(globals())
-del fix_module
